+++ include/Triangle.hh
@@ -14,16 +14,16 @@ namespace Geometry {
         ~Triangle() override = default;
 
         // Operations with triangles
-        double getArea() const override;
-        double getPerimeter() const override;
-        void move(double dx, double dy) override;
-        double getDistance(const Point3D& p1, const Point3D& p2) const;
+        float getArea() const override;
+        float getPerimeter() const override;
+        void move(float dx, float dy) override;
+        float getDistance(const Point3D& p1, const Point3D& p2) const;
 
         // Support function for Barycentric function
-        double TriArea2D(double x1, double y1, double x2, double y2, double x3, double y3);
+        float TriArea2D(float x1, float y1, float x2, float y2, float x3, float y3);
         // Compute barycentric coordinates (u, v, w) for
         // Point3D p with respect to triangle (a, b, c)
-        void Barycentric(Point3D a, Point3D b, Point3D c, Point3D p, double& u, double& v, double& w);
+        void Barycentric(Point3D a, Point3D b, Point3D c, Point3D p, float& u, float& v, float& w);
         // Test if Point3D p is contained in triangle (a, b, c)
         bool TestPoint3DTriangle(Point3D p, Point3D a, Point3D b, Point3D c);
     };
